@@ -1,9 +1,8 @@
 ﻿using System;
 using System.Runtime.CompilerServices;
-using System.Text;
 using System.Threading.Tasks;
 
-[assembly:InternalsVisibleTo("Nanoid.Test")]
+[assembly: InternalsVisibleTo("Nanoid.Test")]
 
 namespace Nanoid
 {
@@ -13,7 +12,8 @@ namespace Nanoid
     public static class Nanoid
     {
         private const string DefaultAlphabet = "_-0123456789abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ";
-        private static readonly CryptoRandom Random = new CryptoRandom();
+        private static readonly CryptoRandom Random = new();
+
         /// <summary>
         /// 
         /// </summary>
@@ -21,13 +21,14 @@ namespace Nanoid
         /// <param name="size"></param>
         /// <returns></returns>
         public static async Task<string> GenerateAsync(string alphabet = DefaultAlphabet, int size = 21) => await Task.Run(() => Generate(Random, alphabet, size));
+
         /// <summary>
         /// 
         /// </summary>
         /// <param name="alphabet"></param>
         /// <param name="size"></param>
         /// <returns></returns>
-        public static string Generate(string alphabet= DefaultAlphabet, int size=21) => Generate(Random, alphabet, size);
+        public static string Generate(string alphabet = DefaultAlphabet, int size = 21) => Generate(Random, alphabet, size);
 
         /// <summary>
         /// 
@@ -38,27 +39,27 @@ namespace Nanoid
         /// <returns></returns>
         /// <exception cref="ArgumentNullException"></exception>
         /// <exception cref="ArgumentOutOfRangeException"></exception>
-        public static string Generate(Random random, string alphabet= DefaultAlphabet, int size=21)
+        public static string Generate(Random random, string alphabet = DefaultAlphabet, int size = 21)
         {
 
             if (random == null)
             {
-                throw new ArgumentNullException("random cannot be null.");
+                throw new ArgumentNullException(nameof(random), "random cannot be null.");
             }
 
             if (alphabet == null)
             {
-                throw new ArgumentNullException("alphabet cannot be null.");
+                throw new ArgumentNullException(nameof(alphabet), "alphabet cannot be null.");
             }
 
             if (alphabet.Length <= 0 || alphabet.Length >= 256)
             {
-                throw new ArgumentOutOfRangeException("alphabet must contain between 1 and 255 symbols.");
+                throw new ArgumentOutOfRangeException(nameof(alphabet), "alphabet must contain between 1 and 255 symbols.");
             }
 
             if (size <= 0)
             {
-                throw new ArgumentOutOfRangeException("size must be greater than zero.");
+                throw new ArgumentOutOfRangeException(nameof(size), "size must be greater than zero.");
             }
 
             // See https://github.com/ai/nanoid/blob/master/format.js for
@@ -67,13 +68,8 @@ namespace Nanoid
             var mask = (2 << 31 - Clz32((alphabet.Length - 1) | 1)) - 1;
             var step = (int)Math.Ceiling(1.6 * mask * size / alphabet.Length);
 
-#if NETSTANDARD2_1
             Span<char> idBuilder = stackalloc char[size];
             Span<byte> bytes = stackalloc byte[step];
-#else
-            var idBuilder = new char[size];
-            var bytes = new byte[step];
-#endif
 
             int cnt = 0;
 
@@ -84,11 +80,13 @@ namespace Nanoid
 
                 for (var i = 0; i < step; i++)
                 {
-
                     var alphabetIndex = bytes[i] & mask;
 
-                    if (alphabetIndex >= alphabet.Length) continue;
+                    if (alphabetIndex >= alphabet.Length) 
+                        continue;
+
                     idBuilder[cnt] = alphabet[alphabetIndex];
+
                     if (++cnt == size)
                     {
                         return new string(idBuilder);
@@ -111,18 +109,21 @@ namespace Nanoid
         internal static int Clz32(int x)
         {
             const int numIntBits = sizeof(int) * 8; //compile time constant
+
             //do the smearing
             x |= x >> 1;
             x |= x >> 2;
             x |= x >> 4;
             x |= x >> 8;
             x |= x >> 16;
+
             //count the ones
             x -= x >> 1 & 0x55555555;
             x = (x >> 2 & 0x33333333) + (x & 0x33333333);
             x = (x >> 4) + x & 0x0f0f0f0f;
             x += x >> 8;
             x += x >> 16;
+
             return numIntBits - (x & 0x0000003f); //subtract # of 1s from 32
         }
     }
